+++ client/app/(todos)/hooks/useGetTodosQuery.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query'
 import { useGetHouseId } from '@/_hooks'
 import { TodoType } from '@/_types'
-import { getTodosApi } from './api'
+import { getTodosApi } from '../api'
 
 export const useGetTodosQuery = () => {
   const { houseId } = useGetHouseId()
